Pantry adam

+++ frontend/src/components/RecipeModal/RecipeToolbar.js
@@ -2,6 +2,7 @@ import React, { Component } from "react";
 import {Button, TextField} from "@material-ui/core";
 import RecipePrinter from "../RecipePrinter/RecipePrinter";
 import { connect } from "react-redux";
+import PlanningModal from './PlanningModal';
 import "./RecipeModal.css";
 import Axios from "axios";
 import AlertDialog from "../AlertDialog/AlertDialog";
@@ -10,15 +11,20 @@ import EditRecipeModal from "../EditRecipeModal/EditRecipeModal";
 class RecipeToolbar extends Component {
   constructor(props) {
     super(props);
-    this.state={
-      a:false,
-      q:false,
+    this.state = {
+      a: false,
+      q: false,
+      p: false,
+      modal:false,
       displayingEditModal: false,
       quantity: 1
     };
-
     this.groceryListAlert = React.createRef();
     this.saveSuccessAlert = React.createRef();
+  }
+
+  closeModal=()=>{
+    this.setState({modal:false});
   }
 
   saveRecipe = () => {
@@ -26,22 +32,24 @@ class RecipeToolbar extends Component {
     Axios.post("http://localhost:8080/savedRecipes", {
       googleId: that.props.auth,
       recipe: that.props.recipe
-    }).then(()=>{
-      that.setState({p:!that.state.p});
+    }).then(() => {
+      that.setState({ p: !that.state.p });
       this.saveSuccessAlert.current.open();
+
+      if (this.props.onSave) this.props.onSave();
+      that.props.handleClose();
     });
   };
 
   removeRecipe = () => {
-    let that=this;
+    let that = this;
     Axios.post("http://localhost:8080/deleteSavedRecipe", {
       googleId: that.props.auth,
       deleteID: that.props.recipe.id
-    }).then(()=>{
-      that.setState({p:!that.state.p});
+    }).then(() => {
+      that.setState({ p: !that.state.p });
 
-      if (this.props.onDelete)
-        this.props.onDelete();
+      if (this.props.onDelete) this.props.onDelete();
     });
   };
 
@@ -85,6 +93,25 @@ class RecipeToolbar extends Component {
     });
   };
 
+  //Fix this when route is implemented
+  removePlanning = () => {
+    // console.log(this.props)
+    let that = this;
+    Axios.post("http://localhost:8080/meal/delete", {
+      googleId: that.props.auth,
+      meal: that.props.meal,
+      date: that.props.date,
+      recipeId:that.props.id,
+    }).then(response => {
+      // console.log(response.data);
+      if(response.data.success){
+        that.props.handleClose();
+      }else{
+        alert('Something went wrong, please try again');
+      }
+    });
+  };
+
   renderSave = () => {
     let that = this;
     let test = null;
@@ -92,7 +119,17 @@ class RecipeToolbar extends Component {
     if (that.props.auth == null || that.props.auth == false) {
       return <div />;
     } else {
-      if (that.props.type === "search" || !that.state.p) {
+      if (that.props.type == "Planning") {
+        test = (
+          <Button
+            variant="contained"
+            color="primary"
+            onClick={that.removePlanning}
+          >
+            Remove from meal plan
+          </Button>
+        );
+      } else if (that.props.type === "search" || !that.state.p) {
         test = (
           <Button variant="contained" color="primary" onClick={that.saveRecipe}>
             Save
@@ -200,7 +237,18 @@ class RecipeToolbar extends Component {
 
         <RecipePrinter recipe={recipe} />
 
-        {this.props.type === 'saved' ? <Button variant="contained" onClick={this.props.saveNote} >Save Notes</Button> : null}
+        {this.props.type === "saved"
+          ? <Button variant="contained" onClick={this.props.saveNote}>
+              Save Notes
+            </Button>
+          : null}
+
+        {this.props.type === "saved"
+          ? <Button variant="contained" onClick={()=>this.setState({modal:true})}>
+              Add to meal plan
+            </Button>
+          : null}
+        <PlanningModal {...this.props} modal={this.state.modal} closeModal={this.closeModal}/>
 
         { quantityInput }
